Move chai-setup.ts to internal directory
