+++ packages/hardhat-toolbox-viem/src/index.ts
@@ -2,7 +2,7 @@ import "@nomicfoundation/hardhat-verify";
 import "@nomicfoundation/hardhat-viem";
 import "hardhat-gas-reporter";
 import "solidity-coverage";
-import "./chai-setup";
+import "./internal/chai-setup";
 
 /**
  * If a new official plugin is added, make sure to update:
